Factor out signature checker (#1700)
